rename gp modules and classes
