+++ autoemulate/emulators/gaussian_process.py
@@ -1,23 +1,45 @@
-import mogp_emulator
-import numpy as np
+from scipy.stats import loguniform
+from scipy.stats import uniform
 from sklearn.base import BaseEstimator
 from sklearn.base import RegressorMixin
+from sklearn.gaussian_process import GaussianProcessRegressor
+from sklearn.gaussian_process.kernels import Matern
+from sklearn.gaussian_process.kernels import RationalQuadratic
+from sklearn.gaussian_process.kernels import RBF
 from sklearn.utils.validation import check_array
 from sklearn.utils.validation import check_is_fitted
 from sklearn.utils.validation import check_X_y
 from skopt.space import Categorical
+from skopt.space import Integer
 from skopt.space import Real
 
+from autoemulate.utils import _suppress_convergence_warnings
 
-class GaussianProcessMOGP(BaseEstimator, RegressorMixin):
+
+class GaussianProcess(BaseEstimator, RegressorMixin):
     """Gaussian Process Emulator.
 
-    Wraps Gaussian Process Regression from the mogp_emulator package.
+    Wraps GaussianProcessRegressor from scikit-learn.
     """
 
-    def __init__(self, nugget="fit"):
+    def __init__(
+        self,
+        kernel=RBF(),
+        alpha=1e-10,
+        optimizer="fmin_l_bfgs_b",
+        n_restarts_optimizer=20,
+        normalize_y=True,
+        copy_X_train=True,
+        random_state=None,
+    ):
         """Initializes a GaussianProcess object."""
-        self.nugget = nugget
+        self.kernel = kernel
+        self.alpha = alpha
+        self.optimizer = optimizer
+        self.n_restarts_optimizer = n_restarts_optimizer
+        self.normalize_y = normalize_y
+        self.copy_X_train = copy_X_train
+        self.random_state = random_state
 
     def fit(self, X, y):
         """Fits the emulator to the data.
@@ -26,7 +48,7 @@ class GaussianProcessMOGP(BaseEstimator, RegressorMixin):
         ----------
         X : {array-like, sparse matrix}, shape (n_samples, n_features)
             The training input samples.
-        y : array-like, shape (n_samples,)
+        y : array-like, shape (n_samples,) or (n_samples, n_outputs)
             The target values (real numbers).
 
         Returns
@@ -34,10 +56,20 @@ class GaussianProcessMOGP(BaseEstimator, RegressorMixin):
         self : object
             Returns self.
         """
-        X, y = check_X_y(X, y, multi_output=False, y_numeric=True)
+
+        X, y = check_X_y(X, y, multi_output=True, y_numeric=True)
         self.n_features_in_ = X.shape[1]
-        self.model_ = mogp_emulator.GaussianProcess(X, y, nugget=self.nugget)
-        self.model_ = mogp_emulator.fit_GP_MAP(self.model_, n_tries=15)
+        self.model_ = GaussianProcessRegressor(
+            kernel=self.kernel,
+            alpha=self.alpha,
+            optimizer=self.optimizer,
+            n_restarts_optimizer=self.n_restarts_optimizer,
+            normalize_y=self.normalize_y,
+            copy_X_train=self.copy_X_train,
+            random_state=self.random_state,
+        )
+        with _suppress_convergence_warnings():
+            self.model_.fit(X, y)
         self.is_fitted_ = True
         return self
 
@@ -60,19 +92,26 @@ class GaussianProcessMOGP(BaseEstimator, RegressorMixin):
         """
         X = check_array(X)
         check_is_fitted(self, "is_fitted_")
-        if return_std:
-            mean, var, _ = self.model_.predict(X)
-            return np.asarray(mean), np.asarray(np.sqrt(var))
-        else:
-            return np.asarray(self.model_.predict(X).mean)
+        return self.model_.predict(X, return_std=return_std)
 
     def get_grid_params(self, search_type="random"):
         """Returns the grid parameters of the emulator."""
         param_space_random = {
-            "nugget": ["fit", "adaptive", "pivot"],
+            "kernel": [
+                RBF(),
+                Matern(),
+                RationalQuadratic(),
+                # DotProduct(),
+            ],
+            "optimizer": ["fmin_l_bfgs_b"],
+            "alpha": loguniform(1e-10, 1e-2),
+            "normalize_y": [True],
         }
         param_space_bayes = {
-            "nugget": Categorical(["fit", "adaptive", "pivot"]),
+            # "kernel": Categorical([RBF(), Matern()]), # unhashable type
+            "optimizer": Categorical(["fmin_l_bfgs_b"]),
+            "alpha": Real(1e-10, 1e-2, prior="log-uniform"),
+            "normalize_y": Categorical([True]),
         }
 
         if search_type == "random":
@@ -84,7 +123,7 @@ class GaussianProcessMOGP(BaseEstimator, RegressorMixin):
 
     @property
     def model_name(self):
-        return "GaussianProcessMOGP"
+        return self.__class__.__name__
 
     def _more_tags(self):
-        return {"multioutput": False}
+        return {"multioutput": True}
